refactor(ai): Move `model/ai.ts` into `~/schema`
Fixes #1594
